refactor(quiz): remove single-part quiz support and legacy components
- Delete deprecated single-part quiz JSON data file
- Remove legacy normalization and conversion utilities for single-part quizzes
- Eliminate related type definitions and type guards for single-part quizzes
- Remove legacy quiz header and footer components and their interfaces
- Simplify quiz data loading to use only multi-part quiz format
- Update imports and type annotations to rely solely on multi-part quiz types
- Refactor question grouping and rendering logic with explicit typing
- Clean up quiz component exports by removing legacy exports
