+++ src/components/quiz/index.ts
@@ -1,9 +1,10 @@
+export { DragDropQuestion } from './drag-drop-question'
 export { default as HighlightManagementPopup } from './highlight-management-popup'
 export { CompactPartSelector, default as PartNavigation, PartNavigationItem } from './part-navigation'
 export { default as QuizCard } from './quiz-card'
 export { default as QuizContentWithSelection } from './quiz-content-with-selection'
-export { LegacyQuizFooter, default as QuizFooter } from './quiz-footer'
-export { LegacyQuizHeader, default as QuizHeader } from './quiz-header'
+export { default as QuizFooter } from './quiz-footer'
+export { default as QuizHeader } from './quiz-header'
 export { default as TextSelectionPopup } from './text-selection-popup'
 export { useQuizState } from './use-quiz-state'
